# WIP - Use the unsafe package to perform direct memory copies during conversion