--- a/hack/update-codegen.sh
+++ b/hack/update-codegen.sh
@@ -47,7 +47,7 @@ ${clientgen} --clientset-name="release_1_3" --input="api/v1,extensions/v1beta1,a
 # Clientgen for federation clientset.
 ${clientgen} --clientset-name=federation_internalclientset --clientset-path=k8s.io/kubernetes/federation/client/clientset_generated --input="../../federation/apis/federation/","api/" --included-types-overrides="api/Service"   "$@"
 ${clientgen} --clientset-name=federation_release_1_3 --clientset-path=k8s.io/kubernetes/federation/client/clientset_generated --input="../../federation/apis/federation/v1alpha1","api/v1" --included-types-overrides="api/v1/Service"   "$@"
-${conversiongen} "$@"
+${conversiongen} --unsafe "$@"
 ${deepcopygen} "$@"
 ${setgen} "$@"