… cmd/libs/go2idl/go-to-protobuf/protobuf/parser.go
… cmd/libs/go2idl/parser/parse_test.go
… federation/apis/federation/v1alpha1/conversion_generated.go
… pkg/api/v1/conversion_generated.go
… pkg/apis/authentication.k8s.io/v1beta1/conversion_generated.go
… pkg/apis/authorization/v1beta1/conversion_generated.go
… pkg/apis/autoscaling/v1/conversion_generated.go
… pkg/apis/batch/v1/conversion_generated.go
… pkg/apis/batch/v2alpha1/conversion_generated.go
… pkg/apis/componentconfig/v1alpha1/conversion_generated.go
… pkg/apis/extensions/v1beta1/conversion_generated.go
… pkg/apis/policy/v1alpha1/conversion_generated.go
… pkg/apis/rbac/v1alpha1/conversion_generated.go
… pkg/apis/rbac/v1alpha1/generated.pb.go
… pkg/apis/rbac/v1alpha1/generated.proto
… pkg/apis/rbac/v1alpha1/types.generated.go
… pkg/apis/rbac/v1alpha1/types.go