The majority of our internal - external conversions are between identical types (memory identical, down to the individual field). This alters the generated conversions to use the unsafe package (optionally) to perform a direct copy of types with identical memory layout and size. The calculation to determine equality is identical types all the way down in the same order. Presence of a custom conversion at any level should terminate the equality check at that level.

Work still remaining:

* [ ] Ensure we have correct testing to catch any issues
* [ ] Add the check that if a conversion or defaulter is registered anywhere down the chain we treat them not the same
* [ ] Identify types that are not optimally aligned and align them
* [ ] Move defaulters up the stack (top level defaulter?)
* [ ] Allow both unsafe and safe versions to be compiled in if necessary so that clients that run where unsafe is not allowed can run

The first five commits are from the optional slices pr #27258

```
benchmark                                      old ns/op     new ns/op     delta
BenchmarkPodConversion-8                       8629          3286          -61.92%
BenchmarkNodeConversion-8                      4986          2056          -58.76%
BenchmarkReplicationControllerConversion-8     4018          3129          -22.13%
BenchmarkEncodeCodecFromInternalProtobuf-8     12765         9040          -29.18%
BenchmarkDecodeCodecToInternalProtobuf-8       19052         14624         -23.24%
BenchmarkEncodeCodecFromInternal-8             43218         39110         -9.51%
BenchmarkDecodeCodec-8                         111268        95846         -13.86%

benchmark                                      old allocs     new allocs     delta
BenchmarkPodConversion-8                       49             12             -75.51%
BenchmarkNodeConversion-8                      18             6              -66.67%
BenchmarkReplicationControllerConversion-8     20             14             -30.00%
BenchmarkEncodeCodecFromInternalProtobuf-8     32             9              -71.88%
BenchmarkDecodeCodecToInternalProtobuf-8       132            109            -17.42%
BenchmarkEncodeCodecFromInternal-8             101            78             -22.77%
BenchmarkDecodeCodec-8                         384            361            -5.99%

benchmark                                      old bytes     new bytes     delta
BenchmarkPodConversion-8                       6338          2446          -61.41%
BenchmarkNodeConversion-8                      4560          1456          -68.07%
BenchmarkReplicationControllerConversion-8     3056          2672          -12.57%
BenchmarkEncodeCodecFromInternalProtobuf-8     7200          4800          -33.33%
BenchmarkDecodeCodecToInternalProtobuf-8       10345         7944          -23.21%
BenchmarkEncodeCodecFromInternal-8             11071         8640          -21.96%
BenchmarkDecodeCodec-8                         19313         16913         -12.43%
```

@kubernetes/sig-api-machinery

<!-- Reviewable:start -->
---
This change is [<img src="https://reviewable.kubernetes.io/review_button.svg" height="34" align="absmiddle" alt="Reviewable"/>](https://reviewable.kubernetes.io/reviews/kubernetes/kubernetes/27659)
<!-- Reviewable:end -->