@smarterclayton 2016-06-18
@k8s-bot test this issue #IGNORE