@smarterclayton 2016-06-18
Built ins / pointers should use the normal path, simpler and less obscure