@wojtek-t 2016-06-20
@smarterclayton - IIUC this is build on top of #27258 - so only last 3 commits are unique, right?