@wojtek-t 2016-07-04
@smarterclayton - can you please rebase? I will look into it then (the underlying PR is already merged).