@smarterclayton 2016-07-05
I think I'm going to separate defaulters from conversion (generate a top level method that invokes all defaulters on a v1 object in the correct order) first.  By doing that, we can do defaulting first, then invoke the conversion, and it should be possible to copy all the way down (i.e. a defaulter deep in Container won't prevent Container from being copied).