@wojtek-t 2016-07-05
@smarterclayton - that makes perfect sense to me. Please ping me when it will be ready for review.