@wojtek-t 2016-07-28
@smarterclayton - what are your plans for this PR? It would be great if we could have this in. Would you find some time soon to finish it?