@smarterclayton 2016-08-01
I think my estimation of the complexity of doing this "right" was under specced.  What needs to happen:

1. Only a type that has no custom conversions can be round tripped this way
2. To make that decision, we need to walk both objects and determine that the conversion can't happen
3. The only place we "know" that today is inside the recursive generation of the methods themselves (since we walk the objects and do the comparisons)
4. To correctly do this optimization, we have to refactor the conversion generation code to do two passes - one pass to calculate the tree of conversions (typeA can be safely cast to typeB because no nested conversions or defaulters exist), and then one pass to turn that tree into the actual methods. 
5. Splitting defaulting is optional but desirable (increases the number of types that can safely be cast this way).