@smarterclayton 2016-08-04
I did get quite a few runs with it, so i'm not as concerned with "default
setup for Kube", but I am for arm, ppc, app engine, etc environments.  We'd
want to do a fair amount of fuzzing, and also ensure that we have a way to
easily disable it for certain platforms.

There was some discussion in golang that many of these assignments should
actually be allowed according to the language spec, but that's fairly far
out.

On Thu, Aug 4, 2016 at 5:28 PM, Daniel Smith <notifications@github.com>
wrote:

> Yeah, I'm a bit nervous about adding our first use of the unsafe package,
> we should be careful.
>
> —
> You are receiving this because you were mentioned.
> Reply to this email directly, view it on GitHub
> <https://github.com/kubernetes/kubernetes/pull/27659#issuecomment-237688972>,
> or mute the thread
> <https://github.com/notifications/unsubscribe-auth/ABG_p1p4PaLzovLEAxGe9KkxDUS7bSy_ks5qcllrgaJpZM4I43ji>
> .
>