@wojtek-t 2016-10-25
This was replaced by #34784 - so closing this one.